fixed [KnightMoveValidator] lossy conversion of int into byte (while performing arthimetic operation)

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/validators/KnightMoveValidator.java
@@ -4,8 +4,8 @@ public class KnightMoveValidator
 public static byte validateMove(byte startRowIndex,byte startColumnIndex,byte destinationRowIndex,byte destinationColumnIndex)
 {
 //mandatoring knight to move in L-shape movement
-byte d1=(startRowIndex<destinationRowIndex)?destinationRowIndex-startRowIndex:startRowIndex-destinationRowIndex;
-byte d2=(startColumnIndex<destinationColumnIndex)?destinationColumnIndex-startColumnIndex:startColumnIndex-destinationColumnIndex;
+byte d1=(byte)((startRowIndex<destinationRowIndex)?destinationRowIndex-startRowIndex:startRowIndex-destinationRowIndex);
+byte d2=(byte)((startColumnIndex<destinationColumnIndex)?destinationColumnIndex-startColumnIndex:startColumnIndex-destinationColumnIndex);
 if(d1!=1 && d1!=2) return 0;
 if(d1==1)//two square right or left and one square bottom or top
 {
